requirements.txt
# Intended for use with Python 3.11

# Development and packaging tools
ipython==8.16.1
setuptools==68.2.2
pre-commit==3.5.0
black==23.10.1
pytest==7.4.3
ruff==0.1.3

# Scientific and numerical libraries
numpy==1.26.1
scipy==1.11.3
pandas==2.1.2
cvxpy==1.4.1

# Visualization libraries
matplotlib==3.8.0
plotly==5.18.0
Pillow==10.0.1

# Streamlit libraries for web applications
streamlit==1.28.0
streamlit-globe==0.2.1
streamlit-extras==0.3.4
